backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django==5.2.1",
    "django-cors-headers==4.7.0",
    "fastapi==0.115.12",
    "httpx==0.27.0",
    "langchain>=0.0.283",
    "langchain-openai>=0.0.5",
    "langgraph>=0.0.18",
    "mermaid-py==0.8.0",
    "python-dotenv==1.0.1",
    "python-multipart==0.0.20",
    "requests==2.32.3",
    "uvicorn==0.34.2",
]
